Split AuthenticationCard into its own folder
